rcamp/rcamp/requirements.txt
Django==1.8.5
django-crispy-forms==1.5.2
django-easy-pjax==1.2.0
django-filter==1.0.2
django-filters==0.2.1
django-grappelli==2.7.2
django-material==0.4.1
djangorestframework==3.3.2
funcparserlib==0.3.6
funcsigs==0.4
mock==1.3.0
mockldap==0.2.6
pbr==1.8.1
python-dateutil==2.5.3
python-ldap==2.4.22
python-pam==1.8.2
pytz==2016.4
six==1.10.0
uWSGI==2.0.17
